Extract no_magic_argument into its own class
