+++ lib/rubocop/cop/magic_numbers/no_magic_argument.rb
@@ -1,0 +1,10 @@
+require_relative "no_magic_numbers"
+
+module RuboCop
+  module Cop
+    module MagicNumbers
+      class NoMagicAssignment < NoMagicNumbers
+      end
+    end
+  end
+end
